Phasing: A second pivot point and better Interface
This commit implements a second pivot point for second order phase
corrections. The idea is based of the discussion in PR#148.
This change allows for two constant phase points during second order
phasing, making it easier to use than before.

Additionally the option for second order phasing is now available
directly in the phasing dialog for better visibility, as discussed in
PR#148. There also exists a persistent option in the preferences
dialog for quicker access of the feature if needed.