fix bug in reading old Bruker spectra (XDIM)
XDIM parameter is not used in 1D processing. In old topspin versions, the
XDIM for 1D is inconsistent with array shape. Trying to to apply the same
reshuffling using XDIM for 1D spectra as for nD failed in some datasets.
The commit check if data is 1D and skip the reshuffing part in that case.