New Feature: Shift Xax when rolling spectrum
Add an option to shift the axis while rolling the data. The option
appears as a Shift Axis checkBox. This makes the peaks to appear at
the same ppm/Hz position after rolling the data.
This is very useful to unfold a 2D spectrum.
The axis shift is implemented by moving the carrier frequency.
This commit closes #155.