Add new apodization slopes for STMAS
Add new apodization slopes for STMAS and use new automatic
calculation of slope.
Calculation principle for R and other ratio implies correlation
of -1Q in t2 with -pQ in t1 with p that depends on MQMAS (3, 5, 7 or 9)
and is 1 for STMAS and 2 for DQMAS

R ratio can be >0 or <0 and corresponds to the slope in frequency domain
-R ratio is to be input to shear the spectrum.

For apodization slope, I chose abs(R) value, but value can be edited
to change sign easily without having to return to user defined drop down.