Bruker import enhanced metadata information
When importing nD Bruker raw data (ser file), metadata will contains whether a dimention
should be converted to hypercomplex or not and the conversion to use from FnMODE parameter.
State, State-TPPI etc.

MASR metadata is only provided for datasets recorded with topspin 4.1.1 and above
as this information was not reliable for earlier versions which could be misleading.