BrukerIO parameter file may contain non ascii char
Reading parameter file may fail if file contains non ascii file.
According to format, Bruker parameter files are JCAMP-DX 5.0 files.
So they should be using only ascii characters. But in some cases
this may not be the case and an exception is raised.
Some text parameter like TI could be filled with non ascii char.
This commit tries to open in ascii first, then if UnicodeDecodeError
is raised, tries in turn utf-8 and latin1 encodings.