Fix/Enhance WindowsRun.bat run options
A bug occurs if %mypath% contains " because the start command assumes
that if the first argument is quotedi "" then it is the label of
the terminal. In that case %curdir% is consumed as a label and ssnake.py
is not found. The fix is simply to define explicitely the terminal label.
For WindowsRun.bat script to work, pythonw or pyw must be in the search path.
During anaconda/miniconda python installation adding path to miniconda in
users environment path is not recommended. This could clash with other python
installations.
An alternative is to define a variable SSNAKE4PYTHON in user environement
variables that points to the required python executable. If such variable
exists it is used in priority, otherwise old behavior is used.